+++ CHANGELOG.md
@@ -8,6 +8,7 @@ New functionality added in a backwards-compatible manner:
 Bug fixes:
 * [OLMIS-4938](https://openlmis.atlassian.net/browse/OLMIS-4938): Moved locked requisition template columns to the top.
 * [OLMIS-4555](https://openlmis.atlassian.net/browse/OLMIS-4555): Add reload flag to reflect changes on admin programs screen.
+* [OLMIS-5134](https://openlmis.atlassian.net/browse/OLMIS-5134): The Approve button will not be visible if a user does not have a right to a supervisory node at which the user wants to approve the requisition.
 
 Improvements:
 * [OLMIS-5115](https://openlmis.atlassian.net/browse/OLMIS-5115): Made the requisition print button disabled when going offline